requirements.txt
spacy>=3.7,<4
pyyaml>=6,<7
g2p-en>=2.1,<3
